requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
execnet==2.0.2
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
pytest-xdist==3.3.1
requests==2.31.0
urllib3==2.0.7
